+++ CHANGELOG.md
@@ -19,3 +19,7 @@
 ## v1.1
 1. As Priced concern
 2. As Pricer concern
+
+## v1.2
+1. Improved the As Priced concern to add the #is_priced! method
+2. Improved the As Pricer concern to add the #price! method